Improve: Precision of normalization
Many distance functions, like Jaccard and Cosine,
have non-trivial normalization procedures at the end.
Like floating-point division or reciprocal square root
approximations.

This commit significantly improves the accuracy of
those operations - `rsqrt` on Skylake and newer now
has accuracy of 2e-14. On older x86 the accuracy is
1.5*2e-12. So there is a 6x error reduction, or up to
36x assuming the product of two such reciprocals.

Moreover, in cases when a zero vector are supplied
to `cos`  as one of the arguments - the distance will be 1.
If both are zero vectors - the distance will be 0.

Closes #153

+++ include/simsimd/types.h
@@ -175,12 +175,12 @@
 
 #if !defined(SIMSIMD_RSQRT)
 #include <math.h>
-#define SIMSIMD_RSQRT(x) (1 / sqrtf(x))
+#define SIMSIMD_RSQRT(x) (1 / sqrt(x))
 #endif
 
 #if !defined(SIMSIMD_LOG)
 #include <math.h>
-#define SIMSIMD_LOG(x) (logf(x))
+#define SIMSIMD_LOG(x) (log(x))
 #endif
 
 #if !defined(SIMSIMD_F32_DIVISION_EPSILON)
